add into_json_map_sized and move into sized/ subfolder
