+++ src/map_iter_to_json.rs
@@ -1,7 +1,9 @@
 
-use std::any::{Any, TypeId};
+use std::any::Any;
 use std::cell::RefCell;
-use serde::ser::{Serialize, Serializer, SerializeMap, Error};
+use serde::ser::{Serialize, Serializer};
+
+use crate::serde_with_utils::serialize_iter_to_map;
 
 /// Blanket impl [to_json_map()](trait.MapIterToJson.html#method.to_json_map) for all `IntoIterator<Item=(&K,&V)>` types.
 pub trait MapIterToJson<'a,K,V>: IntoIterator<Item=(&'a K,&'a V)> where
@@ -69,54 +71,6 @@ V: 'a + Serialize,
 <Self as IntoIterator>::IntoIter: 'a
 { }
 
-pub trait MapIterToJsonSized<'a,K,V>: IntoIterator<Item=(&'a K,&'a V)> where
-Self: Sized,
-K: 'a + Serialize + Any,
-V: 'a + Serialize,
-<Self as IntoIterator>::IntoIter: 'a + ExactSizeIterator
-{
-  fn to_json_map_sized(self) -> Result<String, serde_json::Error> {
-    serde_json::to_string(&SerializeMapIterWrapper {
-      iter: RefCell::new(self.into_iter())
-    })
-  }
-}
-
-impl<'a,K,V,T> MapIterToJsonSized<'a,K,V> for T where
-T: IntoIterator<Item=(&'a K,&'a V)>,
-K: 'a + Serialize + Any,
-V: 'a + Serialize,
-<Self as IntoIterator>::IntoIter: 'a + ExactSizeIterator
-{ }
-
-fn serialize_iter_to_map<'a,S,K,V,I>(serializer: S, iter: &RefCell<I>, len: Option<usize>) -> Result<S::Ok, S::Error>
-where
-I: Iterator<Item=(&'a K,&'a V)>,
-K: 'a + Serialize + Any,
-V: 'a + Serialize,
-S: Serializer
-{
-  let mut iter = iter.borrow_mut();
-  let mut ser_map = serializer.serialize_map(len)?;
-  // handle strings specially so they don't get escaped and wrapped inside another string
-  // compiler seems to be able to optimize this branch away statically
-  if TypeId::of::<K>() == TypeId::of::<String>() {
-    while let Some((k, v)) = iter.next() {
-      let s = (k as &dyn Any).downcast_ref::<String>().ok_or(S::Error::custom("Failed to serialize String as string"))?;
-      ser_map.serialize_entry(s, &v)?;
-    }
-  } else {
-    while let Some((k, v)) = iter.next() {
-      ser_map.serialize_entry(match &serde_json::to_string(&k)
-      {
-        Ok(key_string) => key_string,
-        Err(e) => { return Err(e).map_err(S::Error::custom); }
-      }, &v)?;
-    }
-  }
-  ser_map.end()
-}
-
 pub(crate) struct SerializeMapIterWrapper<'a,K,V,I> where
 I: Iterator<Item=(&'a K,&'a V)>,
 K: 'a,
@@ -134,26 +88,5 @@ impl<'a,K,V,I> Serialize for SerializeMapIterWrapper<'a,K,V,I> where
     S: Serializer
   {
     serialize_iter_to_map(serializer, &self.iter, None)
-  }
-}
-
-pub(crate) struct SerializeMapIterWrapperSized<'a,K,V,I> where
-I: Iterator<Item=(&'a K,&'a V)> + ExactSizeIterator,
-K: 'a,
-V: 'a
-{
-  pub iter: RefCell<I>
-}
-
-impl<'a,K,V,I> Serialize for SerializeMapIterWrapperSized<'a,K,V,I> where
-  I: Iterator<Item=(&'a K,&'a V)> + ExactSizeIterator,
-  K: Serialize + Any,
-  V: Serialize,
-{
-  fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error> where
-    S: Serializer
-  {
-    let len = self.iter.borrow().len();
-    serialize_iter_to_map(serializer, &self.iter, Some(len))
   }
 }
