+++ src/serde_with_utils.rs
@@ -1,8 +1,10 @@
 
 use std::any::{Any, TypeId};
+use std::cell::RefCell;
 use std::marker::PhantomData;
 use serde::de::{Deserialize, Deserializer, Visitor};
 use serde::de::{MapAccess};
+use serde::ser::{Serialize, SerializeMap, Serializer, Error};
 use std::fmt;
 
 // Some utilities required to implement any_key_map / any_key_vec
@@ -93,4 +95,32 @@ pub fn deserialize<'d,D,C,K,V>(deserializer: D) -> Result<C,D::Error> where
   }
   
   deserializer.deserialize_map(Helper(PhantomData))
+}
+
+pub fn serialize_iter_to_map<'a,S,K,V,I>(serializer: S, iter: &RefCell<I>, len: Option<usize>) -> Result<S::Ok, S::Error>
+where
+I: Iterator<Item=(&'a K,&'a V)>,
+K: 'a + Serialize + Any,
+V: 'a + Serialize,
+S: Serializer
+{
+  let mut iter = iter.borrow_mut();
+  let mut ser_map = serializer.serialize_map(len)?;
+  // handle strings specially so they don't get escaped and wrapped inside another string
+  // compiler seems to be able to optimize this branch away statically
+  if TypeId::of::<K>() == TypeId::of::<String>() {
+    while let Some((k, v)) = iter.next() {
+      let s = (k as &dyn Any).downcast_ref::<String>().ok_or(S::Error::custom("Failed to serialize String as string"))?;
+      ser_map.serialize_entry(s, &v)?;
+    }
+  } else {
+    while let Some((k, v)) = iter.next() {
+      ser_map.serialize_entry(match &serde_json::to_string(&k)
+      {
+        Ok(key_string) => key_string,
+        Err(e) => { return Err(e).map_err(S::Error::custom); }
+      }, &v)?;
+    }
+  }
+  ser_map.end()
 }
